Preferences!
You can now set the update interval and whether or not to do it at all in the preferences of the app. I tried making it so that notifications would have customizable sounds and whatnot, but with API 26's notification categories that became a lot more complicated. So I'll do that later.

Oh, and you can't store your private key anymore, because that was stupid.